# Add BatchPreload to decode slabs in parallel and cache